refactor: break out methods from app.py into submodules
