+++ src/aws/osml/model_runner/database/feature_table.py
@@ -1,5 +1,4 @@
 #  Copyright 2023-2024 Amazon.com, Inc. or its affiliates.
-
 import logging
 import random
 import time
