+++ src/aws/osml/model_runner/inference/__init__.py
@@ -7,6 +7,6 @@
 from .detector import Detector
 from .endpoint_factory import FeatureDetectorFactory
 from .feature_selection import FeatureSelector
-from .feature_utils import calculate_processing_bounds, get_source_property
+from .feature_utils import calculate_processing_bounds, get_extents, get_source_property
 from .http_detector import HTTPDetector
 from .sm_detector import SMDetector
