+++ src/aws/osml/model_runner/tile_worker/tile_worker_utils.py
@@ -1,5 +1,7 @@
 #  Copyright 2023-2024 Amazon.com, Inc. or its affiliates.
 
+import ast
+import json
 import logging
 import tempfile
 from pathlib import Path
@@ -10,19 +12,25 @@ from typing import List, Optional, Tuple
 from aws_embedded_metrics import MetricsLogger
 from aws_embedded_metrics.metric_scope import metric_scope
 from aws_embedded_metrics.unit import Unit
+from geojson import Feature
 from osgeo import gdal
 
 from aws.osml.features import Geolocator, ImagedFeaturePropertyAccessor
 from aws.osml.gdal import GDALConfigEnv
 from aws.osml.image_processing.gdal_tile_factory import GDALTileFactory
-from aws.osml.model_runner.database import RegionRequestItem, RegionRequestTable
+from aws.osml.model_runner.api import RegionRequest
+from aws.osml.model_runner.app_config import MetricLabels, ServiceConfig
+from aws.osml.model_runner.common import (
+    FeatureDistillationDeserializer,
+    ImageRegion,
+    Timer,
+    get_credentials_for_assumed_role,
+)
+from aws.osml.model_runner.database import FeatureTable, RegionRequestItem, RegionRequestTable
+from aws.osml.model_runner.inference import FeatureSelector
+from aws.osml.model_runner.inference.endpoint_factory import FeatureDetectorFactory
 from aws.osml.photogrammetry import ElevationModel, SensorModel
 
-from ..api import RegionRequest
-from ..app_config import MetricLabels, ServiceConfig
-from ..common import Timer, get_credentials_for_assumed_role
-from ..database import FeatureTable
-from ..inference.endpoint_factory import FeatureDetectorFactory
 from .exceptions import ProcessTilesException, SetupTileWorkersException
 from .tile_worker import TileWorker
 from .tiling_strategy import TilingStrategy
@@ -282,3 +290,51 @@ def sizeof_fmt(num: float, suffix: str = "B") -> str:
             return "%3.1f%s%s" % (num, unit, suffix)
         num /= 1024.0
     return "%.1f%s%s" % (num, "Yi", suffix)
+
+
+def select_features(
+    feature_distillation_option: str,
+    features: List[Feature],
+    processing_bounds: ImageRegion,
+    region_size: str,
+    tile_size: str,
+    tile_overlap: str,
+    tiling_strategy: TilingStrategy,
+) -> List[Feature]:
+    """
+    Selects the desired features using the options in the JobItem (NMS, SOFT_NMS, etc.).
+    This code applies a feature selector only to the features that came from regions of the image
+    that were processed multiple times. First features are grouped based on the region they were
+    processed in. Any features found in the overlap area between regions are run through the
+    FeatureSelector. If they were not part of an overlap area between regions, they will be grouped
+    based on tile boundaries. Any features that fall into the overlap of adjacent tiles are filtered
+    by the FeatureSelector. All other features should not be duplicates; they are added to the result
+    without additional filtering.
+
+    Computationally, this implements two critical factors that lower the overall processing time for the
+    O(N^2) selection algorithms. First, it will filter out the majority of features that couldn't possibly
+    have duplicates generated by our tiled image processing; Second, it runs the selection algorithms
+    incrementally on much smaller groups of features.
+
+    :param region_size:
+    :param feature_distillation_option: str = the options used in selecting features (e.g., NMS/SOFT_NMS, thresholds)
+    :param features: List[Feature] = the list of geojson features to process
+    :param processing_bounds: the requested area of the image
+    :param region_size: str = region size to use for feature dedup
+    :param tile_size: str = size of the tiles used during processing
+    :param tile_overlap: str = overlap between tiles during processing
+    :param tiling_strategy: the tiling strategy to use for feature dedup
+    :return: List[Feature] = the list of geojson features after processing
+    """
+    feature_distillation_option_dict = json.loads(feature_distillation_option)
+    feature_distillation_option = FeatureDistillationDeserializer().deserialize(feature_distillation_option_dict)
+    feature_selector = FeatureSelector(feature_distillation_option)
+
+    region_size = ast.literal_eval(region_size)
+    tile_size = ast.literal_eval(tile_size)
+    overlap = ast.literal_eval(tile_overlap)
+    deduped_features = tiling_strategy.cleanup_duplicate_features(
+        processing_bounds, region_size, tile_size, overlap, features, feature_selector
+    )
+
+    return deduped_features
